requirements.txt
aiotools
--editable ./async-solipsism
pytest-aiohttp
pytest-mock
